chg to int

+++ game.cl
@@ -1,4 +1,4 @@
-__kernel void game(__global const float *a, __global const float *b, __global float *c)
+__kernel void game(__global const int *a, __global const int *b, __global int *c)
 {
     int gid = get_global_id(0);
     c[gid] = a[gid] + b[gid];
